catalog: move into its own dir
