+++ src/trx/game/catalog/lara_anims.def
