+++ src/trx/game/catalog/lara_states.def
