+++ src/trx/game/catalog/manager.h
