+++ src/trx/game/catalog/objects.def
