+++ src/trx/game/inventory_ring/vars.c
@@ -1,7 +1,7 @@
 #include <trx/game/inventory_ring/vars.h>
 
 #include <trx/debug.h>
-#include <trx/game/catalog.h>
+#include <trx/game/catalog/manager.h>
 #include <trx/game/shell.h>
 #include <trx/json/util/file.h>
 #include <trx/memory.h>
