+++ src/trx/game/items/actions/ids.h
@@ -9,7 +9,7 @@ typedef enum {
 typedef enum {
     ITEM_TRX_ACTION_INVALID = -1,
 #define X_CATALOG_ID(enum_value) enum_value,
-#include <trx/game/catalog_item_actions.def>
+#include <trx/game/catalog/item_actions.def>
 #undef X_CATALOG_ID
     ITEM_ACTION_NUMBER_OF,
 } ITEM_TRX_ACTION;
