+++ src/trx/game/lara/skin/storage.c
@@ -3,7 +3,7 @@
 #include <trx/config.h>
 #include <trx/debug.h>
 #include <trx/enum_map.h>
-#include <trx/game/catalog.h>
+#include <trx/game/catalog/manager.h>
 #include <trx/game/game_strings/entries.h>
 #include <trx/game/lara.h>
 #include <trx/game/shell.h>
