+++ src/trx/game/shell/flow.c
@@ -1,7 +1,7 @@
 #include <trx/config.h>
 #include <trx/debug.h>
 #include <trx/enum_map.h>
-#include <trx/game/catalog.h>
+#include <trx/game/catalog/manager.h>
 #include <trx/game/clock.h>
 #include <trx/game/console.h>
 #include <trx/game/events.h>
