+++ src/trx/game/sound/ids.c
@@ -1,4 +1,4 @@
-#include <trx/game/catalog.h>
+#include <trx/game/catalog/manager.h>
 #include <trx/game/sound.h>
 
 SAMPLE_ID Sound_ToGameID(const SAMPLE_TRX_ID trx_id)
